Moved RadioButton component to RadioButtonGroup
