+++ src/components/forms/radioButtonGroup/RadioButtonGroup.tsx
@@ -1,12 +1,22 @@
 import React from "react";
 import styles from "src/components/forms/radioButtonGroup/radioButtonGroup.module.css";
-import { RadioButton } from "./components/RadioButton";
 import textStyles from "src/components/text/text.module.css";
 
 export interface IOption {
   id: string;
   label: string;
   disabled?: boolean;
+}
+
+interface RadioButtonProps {
+  id: string;
+  name: string;
+  label: string;
+  value?: string;
+  checked?: boolean;
+  disabled?: boolean;
+  defaultChecked?: boolean;
+  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
 interface RadioButtonGroupProps {
@@ -51,5 +61,39 @@ export const RadioButtonGroup = ({
         ))}
       </div>
     </fieldset>
+  );
+};
+
+const RadioButton = ({
+  id,
+  name,
+  value,
+  label,
+  checked,
+  disabled,
+  defaultChecked,
+  onChange,
+}: RadioButtonProps) => {
+  return (
+    <label
+      htmlFor={id}
+      className={`${styles.container} ${textStyles.caption} ${disabled ? styles.disabledLabel : styles.label}`}
+    >
+      <input
+        className={styles.input}
+        type="radio"
+        name={name}
+        id={id}
+        value={value}
+        checked={checked}
+        disabled={disabled}
+        aria-disabled={disabled ? "true" : "false"}
+        defaultChecked={defaultChecked}
+        onChange={onChange}
+        aria-label={label}
+      />
+
+      {label}
+    </label>
   );
 };
